Ensure we load data as float32

+++ fm4ar/datasets/vasist_2023/__init__.py
@@ -26,9 +26,9 @@ def load_vasist_2023_dataset(config: dict) -> ArDataset:
     dataset_dir = get_datasets_dir() / "vasist-2023" / which
     file_path = dataset_dir / file_name
     with h5py.File(file_path, "r") as hdf_file:
-        theta = np.array(hdf_file["theta"][:n_samples])
-        flux = np.array(hdf_file["spectra"][:n_samples])
-        wlen = np.array(hdf_file["wavelengths"])
+        theta = np.array(hdf_file["theta"][:n_samples], dtype=np.float32)
+        flux = np.array(hdf_file["spectra"][:n_samples], dtype=np.float32)
+        wlen = np.array(hdf_file["wavelengths"], dtype=np.float32)
 
     # Define noise levels
     noise_levels = 1.25754e-17 * 1e16
